Tidy

+++ src/bodypart.h
@@ -351,7 +351,7 @@ class body_part_set
         std::vector<bodypart_str_id> values() const {
             std::vector<bodypart_str_id> vec;
             vec.reserve( parts.count() );
-            for( int i = 0; i < parts.size(); i++ ) {
+            for( size_t i = 0; i < parts.size(); i++ ) {
                 if( parts[i] ) {
                     vec.push_back( bodyparts_str_ids()[ i ] );
                 }
